requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
supabase==2.3.0
pydantic==2.9.1
python-dotenv==1.0.1
requests==2.32.3
